Refactor : ResponseDto단  변수 타입 변경
